사용하지 않는 메서드 삭제
사용하지 않는 주석 제거
Repository 사용하지 않는 쿼리메소드 삭제
ResumeService이름 변경
