+++ ziczone/src/main/java/org/zerock/ziczone/repository/application/CareerRepository.java
@@ -10,9 +10,6 @@ import java.util.List;
 public interface CareerRepository extends JpaRepository<Career, Long> {
     List<Career> findByResume_ResumeId(Long resumeId);
 
-    void deleteByResume_ResumeId(Long resumeId);
-
-    void deleteByResumePersonalUserPersonalId(Long personalUserId);
 
     List<Career>  findByResume(Resume resume);
