+++ ziczone/src/main/java/org/zerock/ziczone/repository/job/JobPositionRepository.java
@@ -13,15 +13,5 @@ public interface JobPositionRepository extends JpaRepository<JobPosition, Long> 
 
     List<JobPosition> findByPersonalUser(PersonalUser personalUser);
 
-    /*
-     Spring Data JPA는 연관된 엔티티의 속성을 탐색할 때 중간에 언더스코어(_)가 필요합니다.
-     여기서 PersonalUser는 JobPosition 또는 TechStack 엔티티와 연관된 엔티티입니다.
-     *findByPersonalUser_PersonalId**는 JobPosition 또는 TechStack 엔티티에서 PersonalUser 엔티티의 personalId 속성을 기준으로 데이터를 조회합니다.
-     즉, JobPosition이나 TechStack 엔티티가 PersonalUser 엔티티와 연관되어 있고, 이 PersonalUser 엔티티의 personalId를 이용하여 데이터를 가져옵니다.
-     */
-    List<JobPosition> findByPersonalUser_PersonalId(Long personalId);
-
-    void deleteByPersonalUser(PersonalUser personalUser);
-
     void deleteByPersonalUserPersonalId(Long personalUserId);
 }
